fix: card page breaking
